refactor GlobOpt::OptArraySrc into helper class
This commit doesn't try to consolidate logic, but makes it easier to do so later, and should make the function easier to deal with in general.
